requirements.txt
dash
psutil
speedtest-cli
plotly
